remove unused tests, streamline 4337 contracts setup in tests
