+++ packages/contracts-account-abstraction/src/setupERC4337Contracts.test.ts
@@ -1,9 +1,12 @@
-import { expect, describe, test, beforeAll } from "vitest";
+import { expect, describe, test } from "vitest";
 import { createPublicClient, createWalletClient, http, nonceManager } from "viem";
 import { localhost } from "viem/chains";
-import { getOrDeployDeterministicDeployer, getLocalAccount } from "@owlprotocol/viem-utils";
+import { entryPoint07Address } from "viem/account-abstraction";
+import { getLocalAccount } from "@owlprotocol/viem-utils";
+
 import { port } from "./test/constants.js";
 import { setupERC4337Contracts, setupVerifyingPaymaster } from "./setupERC4337Contracts.js";
+import { SIMPLE_ACCOUNT_FACTORY_ADDRESS } from "./constants.js";
 
 describe("setupERC4337Contracts.test.ts", function () {
     const chain = {
@@ -25,21 +28,17 @@ describe("setupERC4337Contracts.test.ts", function () {
         transport,
     });
 
-    beforeAll(async () => {
-        //Deploy Deterministic Deployer first
-        const { hash } = await getOrDeployDeterministicDeployer(walletClient);
-        if (hash) {
-            await publicClient.waitForTransactionReceipt({ hash });
-        }
-    });
-
     test("setupERC4337Contracts", async () => {
+        // Note: Contracts are actually deployed in vitest.setup.ts (shared across all tests)
         const result = await setupERC4337Contracts(walletClient);
 
         expect(await publicClient.getCode({ address: result.entrypoint.address })).toBeDefined();
         expect(await publicClient.getCode({ address: result.simpleAccountFactory.address })).toBeDefined();
         expect(await publicClient.getCode({ address: result.entrypointSimulations.address })).toBeDefined();
         expect(await publicClient.getCode({ address: result.pimlicoEntrypointSimulations.address })).toBeDefined();
+
+        expect(result.entrypoint.address).toBe(entryPoint07Address);
+        expect(result.simpleAccountFactory.address).toBe(SIMPLE_ACCOUNT_FACTORY_ADDRESS);
 
         const verifyingPaymaster = await setupVerifyingPaymaster(walletClient, {
             verifyingSignerAddress: walletClient.account.address,
